added Apply feature to settings
