--- include/settings/CustomSettingNode.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <Geode.hpp>
-
-namespace geode {
-    // java
-    using CustomSettingNodeGenerator = std::function<TableViewCell*(Setting*, float)>;
-}
