+++ src/gui/mods/info/ModSettingsLayer.cpp
@@ -18,6 +18,7 @@ bool ModSettingsLayer::init(Mod* mod) {
     m_mainLayer->addChild(bg);
 
     m_buttonMenu = CCMenu::create();
+    m_buttonMenu->setZOrder(10);
     m_mainLayer->addChild(m_buttonMenu);
 
     CCDirector::sharedDirector()->getTouchDispatcher()->incrementForcePrio(2);
@@ -43,10 +44,10 @@ bool ModSettingsLayer::init(Mod* mod) {
     bgSprite->setPosition(winSize.width / 2, winSize.height / 2);
     m_mainLayer->addChild(bgSprite);
 
-	auto settings = ModSettingsList::create(mod, listSize.width, listSize.height);
-	settings->setPosition(winSize.width / 2 - listSize.width / 2, winSize.height / 2 - listSize.height / 2);
-	m_mainLayer->addChild(settings);
-        
+	m_list = ModSettingsList::create(mod, this, listSize.width, listSize.height);
+	m_list->setPosition(winSize.width / 2 - listSize.width / 2, winSize.height / 2 - listSize.height / 2);
+	m_mainLayer->addChild(m_list);
+
     {
         auto topSprite = CCSprite::createWithSpriteFrameName("GJ_commentTop_001.png");
         topSprite->setPosition({ winSize.width / 2, winSize.height / 2 + listSize.height / 2 - 5.f });
@@ -69,6 +70,17 @@ bool ModSettingsLayer::init(Mod* mod) {
         this->addChild(rightSprite);
     }
 
+    m_applyButtonSpr = ButtonSprite::create("Apply", "goldFont.fnt", "GJ_button_01.png", .8f);
+    m_applyButtonSpr->setScale(.8f);
+
+    auto applyBtn = CCMenuItemSpriteExtra::create(
+        m_applyButtonSpr,
+        this,
+        menu_selector(ModSettingsLayer::onApply)
+    );
+    applyBtn->setPosition(size.width / 2 - 80.f, -size.height / 2 + 25.f);
+    m_buttonMenu->addChild(applyBtn);
+
     auto closeSpr = CCSprite::createWithSpriteFrameName("GJ_closeBtn_001.png");
     closeSpr->setScale(.8f);
 
@@ -83,7 +95,37 @@ bool ModSettingsLayer::init(Mod* mod) {
     this->setKeypadEnabled(true);
     this->setTouchEnabled(true);
 
+    this->updateState();
+
     return true;
+}
+
+void ModSettingsLayer::updateState() {
+    if (m_list->hasUnsavedModifiedSettings()) {
+        m_applyButtonSpr->setOpacity(255);
+        m_applyButtonSpr->setColor({ 255, 255, 255 });
+    } else {
+        m_applyButtonSpr->setOpacity(155);
+        m_applyButtonSpr->setColor({ 155, 155, 155 });
+    }
+}
+
+void ModSettingsLayer::FLAlert_Clicked(FLAlertLayer*, bool btn2) {
+    if (btn2) {
+        this->close();
+    }
+}
+
+void ModSettingsLayer::onApply(CCObject*) {
+    if (!m_list->hasUnsavedModifiedSettings()) {
+        FLAlertLayer::create(
+            "No Changes",
+            "No changes were made.",
+            "OK"
+        )->show();
+    }
+    m_list->applyChanges();
+    this->updateState();
 }
 
 void ModSettingsLayer::keyDown(enumKeyCodes key) {
@@ -95,9 +137,21 @@ void ModSettingsLayer::keyDown(enumKeyCodes key) {
 }
 
 void ModSettingsLayer::onClose(CCObject*) {
+    if (m_list->hasUnsavedModifiedSettings()) {
+        return FLAlertLayer::create(
+            this, "Unsaved Changes",
+            "You have <cy>unsaved</c> settings! Are you sure "
+            "you want to <cr>discard</c> your changes?",
+            "Cancel", "Discard", 400.f
+        )->show();
+    }
+    this->close();
+};
+
+void ModSettingsLayer::close() {
     this->setKeyboardEnabled(false);
     this->removeFromParentAndCleanup(true);
-};
+}
 
 ModSettingsLayer* ModSettingsLayer::create(Mod* mod) {
     auto ret = new ModSettingsLayer();
