+++ src/gui/mods/settings/ModSettingsList.hpp
@@ -2,18 +2,27 @@
 
 #include <Geode.hpp>
 #include <nodes/ScrollLayer.hpp>
+#include <settings/SettingNode.hpp>
 
 USE_GEODE_NAMESPACE();
+
+class ModSettingsLayer;
 
 class ModSettingsList : public CCLayer {
 protected:
     Mod* m_mod;
+    ModSettingsLayer* m_settingsLayer;
     ScrollLayer* m_scrollLayer;
+    std::vector<SettingNode*> m_settingNodes;
 
-    bool init(Mod* mod, float width, float height);
+    bool init(Mod* mod, ModSettingsLayer* layer, float width, float height);
 
 public:
     static ModSettingsList* create(
-        Mod* Mod, float width, float height
+        Mod* mod, ModSettingsLayer* layer, float width, float height
     );
+
+    void updateList();
+    bool hasUnsavedModifiedSettings() const;
+    void applyChanges();
 };
